reorganizing a bit
