+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@vdcode/mockit",
-  "version": "1.0.0-beta.4",
+  "version": "1.0.0-beta.5",
   "description": "mocking library experiment",
   "main": "dist/index.js",
   "module": "dist/index.mjs",
