+++ src/classMocks/GetClassMethods.ts
