+++ src/classMocks/index.ts
@@ -1,0 +1,4 @@
+export { AbstractClassMock } from "./abstract";
+export { ConcreteClassMock } from "./concrete";
+
+export type { AbstractClass, Class } from "./types";
