+++ src/classMocks/types.ts
@@ -1,0 +1,2 @@
+export type AbstractClass<T> = abstract new (...args: any[]) => T;
+export type Class<T> = new (...args: any[]) => T;
