+++ src/index.ts
@@ -1,2 +1,11 @@
-import { Mockit } from "./mockit";
-export { Mockit };
+import {
+  Mockit,
+  mock,
+  mockAbstract,
+  mockFunction,
+  spy,
+  suppose,
+  verify,
+  when,
+} from "./mockit";
+export { Mockit, mock, mockAbstract, mockFunction, spy, suppose, verify, when };
