+++ src/internal/functionMock/behaviour.ts
