+++ src/internal/index.ts
@@ -1,0 +1,5 @@
+export { FunctionMock } from "./functionMock";
+export { FunctionMockUtils } from "./functionMock/utils";
+export { FunctionSpy } from "./functionSpy";
+
+export { Behaviour } from "./functionMock/behaviour";
